create API routes with Postgres db, reorg directory to separate concerns
