+++ README.md
@@ -23,9 +23,12 @@ The ratings and history module supports CRUD operations for the two main data so
 | --- | --- | --- | --- | 
 | GET | api/:stockTicker | Retrieve a snapshot of analyst reviews with percentage recommendations to buy, hold or sell for specified stock | {"symbol" : "MSFT", "recBuy" : 7, "recHold" : 6, "recSell" : 15, "reviewBuy" : "Rubber scalable disintermediate markets Microsoft Corporation user-facing Tasty. \n The strategic revolutionize functionalities Quae ea quisquam.. \n Overall, scalable disintermediate markets Microsoft Corporation holistic Voluptatem reiciendis commodi.", "reviewSell" : "Rubber one-to-one repurpose communities user-facing Microsoft Corporation Tasty. \n For intuitive mesh communities Quae ea quisquam.. \n Hence, one-to-one repurpose communities Microsoft Corporation holistic Voluptatem reiciendis commodi."}
 
+| Post | api/stock/:stockID/ratings | Update analyst snapshot when new reviews change the percentage of buy, sell, or hold recommendations | {"symbol" : "ORCL", "purchase_id" : 11478, "name" : "Oracle Corporation", "timeinforce" : "Good for day", "submitted" : ISODate("2015-05-01T11:55:13.029Z"), "status" : "filled", "enteredQuantity" : 50, "filled" : ISODate("2018-03-30T00:50:14.922Z"), "filledQuantityShares" : 50, "filledQuantityPrice" : 50, "total" : 129 }
+
 ### Purchase History
 
 | Method | Endpoint | Description | Result
 | --- | --- | --- | --- | 
 | GET | api/stock/:stockID/history | Retrieve a record of the user’s past purchases of the specified stock | {"symbol" : "ORCL", "purchase_id" : 11478, "name" : "Oracle Corporation", "timeinforce" : "Good for day", "submitted" : ISODate("2015-05-01T11:55:13.029Z"), "status" : "filled", "enteredQuantity" : 50, "filled" : ISODate("2018-03-30T00:50:14.922Z"), "filledQuantityShares" : 50, "filledQuantityPrice" : 50, "total" : 129 }
+
