--- controller.js
@@ -1,5 +1,0 @@
-const { requestStock, requestPurchase } = require('./model');
-
-module.exports.getStocks = stockID => requestStock(stockID);
-
-module.exports.getPurchases = stockID => requestPurchase(stockID);
