+++ database-mongodb/Stock.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-const db = require('./index.js');
+const db = require('../server/index.js');
 
 mongoose.Promise = global.Promise;
