--- mockRequestDB.js
@@ -1,1 +1,0 @@
-//  for jest test
